fix login sprybook error the rght way
