+++ src/App.js
@@ -6,7 +6,7 @@ import { Profile } from './pages/profile.jsx';
 import { Notifications } from './pages/notifications.jsx';
 // import { Friends } from './pages/friends.jsx';
 import { UserContextProvider } from './context/userContext';
-import { LoginForm } from './components/LoginForm/index';
+import { LoginForm } from './pages/login';
 import { ProtectedRoute } from './pages/protected-route';
 import { Post } from './pages/post';
 function App() {
